Move AspectRatioDropdown to Image Editor components
